+++ Ecommerce System GUI/src/main/java/EcommerceSystem/AccountManager.java
@@ -1,8 +1,7 @@
-package EcommerceSystemimport;
+package EcommerceSystem;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Scanner;
 
 public class AccountManager {
     private static List<Account> accounts=new ArrayList<>();
