+++ Ecommerce System GUI/src/main/java/EcommerceSystem/CustomerAccount.java
@@ -1,4 +1,4 @@
-package EcommerceSystemimport;
+package EcommerceSystem;
 
 class CustomerAccount extends Account {
     public CustomerAccount(String username, String password) {
