+++ Ecommerce System GUI/src/main/java/EcommerceSystem/ManagerAccount.java
@@ -1,4 +1,4 @@
-package EcommerceSystemimport;
+package EcommerceSystem;
 
 class ManagerAccount extends Account {
     public ManagerAccount(String username, String password) {
