+++ Ecommerce System GUI/src/main/java/EcommerceSystem/Orders.java
@@ -1,5 +1,4 @@
 package EcommerceSystem;
-
 public class Orders {
     private String orderId;
     private String userEmail;
