+++ Ecommerce System GUI/src/main/java/EcommerceSystem/PaymentProcessor.java
@@ -1,5 +1,4 @@
 package EcommerceSystem;
-
 public class PaymentProcessor {
     private boolean paymentGatewayConnected;
