+++ Ecommerce System GUI/src/main/java/EcommerceSystem/Product.java
@@ -36,4 +36,8 @@ public class Product {
     public void setPrice(double price) {
         this.price = price;
     }
+    @Override
+    public String toString() {
+        return "ID: " + id + ", Name: " + name + ", Price: " + price;
+    }
 }
