+++ Ecommerce System GUI/src/main/java/EcommerceSystem/UserAccount.java
@@ -1,29 +1,40 @@
 package EcommerceSystem;
 
+
 public class UserAccount {
     private String username;
     private String password;
     private boolean loggedIn;
+    private AccountManager accountManager;
 
-    public UserAccount(String username, String password) {
-        this.username = username;
-        this.password = password;
+    public UserAccount(AccountManager accountManager) {
+        this.accountManager = accountManager;
         this.loggedIn = false;
     }
 
     public boolean login(String username, String password) {
-        if (this.username.equals(username) && this.password.equals(password)) {
-            loggedIn = true;
+        if (accountManager.authenticate(username, password)) {
+            this.username = username;
+            this.password = password;
+            this.loggedIn = true;
             return true;
         }
         return false;
     }
 
     public void logout() {
-        loggedIn = false;
+        this.loggedIn = false;
     }
 
     public boolean isLoggedIn() {
         return loggedIn;
     }
+
+    public boolean isManager() {
+        if (isLoggedIn()) {
+            return accountManager.isManager(username);
+        }
+        return false;
+    }
+
 }
