--- Ecommerce System GUI/src/main/java/EcommerceSystemimport/Orders.java
@@ -1,28 +1,0 @@
-package EcommerceSystemimport;
-public class Orders {
-    private String orderId;
-    private String userEmail;
-
-    public Orders(String orderId, String userEmail) {
-        this.orderId = orderId;
-        this.userEmail = userEmail;
-    }
-
-    // Getters and setters
-
-    public String getOrderId() {
-        return orderId;
-    }
-
-    public void setOrderId(String orderId) {
-        this.orderId = orderId;
-    }
-
-    public String getUserEmail() {
-        return userEmail;
-    }
-
-    public void setUserEmail(String userEmail) {
-        this.userEmail = userEmail;
-    }
-}
