--- Ecommerce System GUI/src/main/java/EcommerceSystemimport/Product.java
@@ -1,43 +1,0 @@
-package EcommerceSystemimport;
-
-public class Product {
-    private String id;
-    private String name;
-    private double price;
-
-    public Product(String id, String name, double price) {
-        this.id = id;
-        this.name = name;
-        this.price = price;
-    }
-
-    // Getters and setters
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public double getPrice() {
-        return price;
-    }
-
-    public void setPrice(double price) {
-        this.price = price;
-    }
-    @Override
-    public String toString() {
-        return "ID: " + id + ", Name: " + name + ", Price: " + price;
-    }
-}
