--- Ecommerce System GUI/src/main/java/EcommerceSystemimport/ShoppingCart.java
@@ -1,32 +1,0 @@
-package EcommerceSystemimport;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class ShoppingCart {
-    private List<Product> items;
-
-    public ShoppingCart() {
-        this.items = new ArrayList<>();
-    }
-
-    public void addItem(Product item) {
-        items.add(item);
-    }
-
-    public void removeItem(Product item) {
-        items.remove(item);
-    }
-
-    public List<Product> getItems() {
-        return items;
-    }
-
-    public double calculateTotal() {
-        double total = 0.0;
-        for (Product item : items) {
-            total += item.getPrice();
-        }
-        return total;
-    }
-}
