+++ Ecommerce System GUI/src/main/java/ecommerce/ecommercesystemgui/CreateAccountController.java
@@ -1,6 +1,6 @@
 package ecommerce.ecommercesystemgui;
 
-import EcommerceSystemimport.AccountManager;
+import EcommerceSystem.AccountManager;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.TextField;
@@ -8,13 +8,10 @@ import javafx.stage.Stage;
 import java.io.IOException;
 
 import javafx.event.ActionEvent;
-import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Node;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
-import javafx.scene.control.TextField;
-import javafx.stage.Stage;
 
 public class CreateAccountController {
     private Stage stage;
