+++ Ecommerce System GUI/src/main/java/ecommerce/ecommercesystemgui/LoginController.java
@@ -1,7 +1,7 @@
 package ecommerce.ecommercesystemgui;
 
-import EcommerceSystemimport.AccountManager;
-import EcommerceSystemimport.UserAccount;
+import EcommerceSystem.AccountManager;
+import EcommerceSystem.UserAccount;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.TextField;
@@ -9,13 +9,10 @@ import javafx.stage.Stage;
 import java.io.IOException;
 
 import javafx.event.ActionEvent;
-import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Node;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
-import javafx.scene.control.TextField;
-import javafx.stage.Stage;
 
 public class LoginController {
